requirements.txt
aiogram
httpx
python-dotenv
aiosqlite
requests